requirements.txt
python-dotenv
# openai  # OpenAI 사용 시 주석 해제
anthropic
pyupbit
ta
selenium
webdriver-manager
pillow
youtube-transcript-api
streamlit
plotly
schedule
googletrans==4.0.0-rc1
supabase

# US Stock Trader 추가 패키지
yfinance
pandas
numpy
pydantic
pytz
requests